begin a Mears 2014 process
